review: narrow Cycle 800 audit packet
