+++ docs/audit/data/citation_graph_manifest.json
@@ -4466,6 +4466,10 @@
    "deps_hash": "e3b0c44298fc",
    "out_degree": 0
   },
+  "factorization_object_api_support_note_2026-07-28": {
+   "deps_hash": "759285fc1e81",
+   "out_degree": 1
+  },
   "family_companion_compare_note": {
    "deps_hash": "d65483554fad",
    "out_degree": 3
